Extend low memory range reserved for logs
Some platforms with higher core count had the issue of logs
overflowing the memory range reserved for them, and thus
overwriting themselves, leaving just a bit of last lines
of logs to be later read.

Before range was 32KB, now it is 200KB which HOPEFULLY
won't need further extensions

WARNING:
This patch causes logs to be unreadable with older versions
of txt-stat. If you're using TBOOT with this patch, please
swap your txt-stat for the one built with this patch too.

+++ include/tboot.h
@@ -131,7 +131,7 @@ typedef struct __packed {
 #define ZIP_COUNT_MAX 10
 typedef struct {
     uuid_t     uuid;
-    uint16_t   max_size;
+    uint32_t   max_size;
     uint16_t   curr_pos;
     uint16_t   zip_pos[ZIP_COUNT_MAX];
     uint16_t   zip_size[ZIP_COUNT_MAX];
